Refactor screen imports in Navigator.tsx
