+++ src/components/nav/Navigator.tsx
@@ -39,9 +39,9 @@ import MnemonicScreen from '@screens/Restore/Mnemonic'
 import RecoverScreen from '@screens/Restore/Recover'
 import RecoveringScreen from '@screens/Restore/Recovering'
 import RestoreWarningScreen from '@screens/Restore/RestoreWarning'
+import SeedScreen from '@screens/Restore/Seed'
+import SeedUpdateScreen from '@screens/Restore/SeedUpdate'
 import SelectRecoveryMintScreen from '@screens/Restore/SelectRecoveryMint'
-import SeedScreen from '@screens/Seed'
-import SeedUpdateScreen from '@screens/SeedUpdate'
 import Settings from '@screens/Settings'
 import AboutSettings from '@screens/Settings/About'
 import BackupPage from '@screens/Settings/Backup'
