+++ src/screens/Restore/Seed.tsx
